refactor: abstract handlers and selectors
