+++ src/views/sent_sms_list/handlers.js
@@ -1,0 +1,22 @@
+export const createToggleHandler = ({ selected, setSelected }) => id => {
+    const isSelected = selected.includes(id)
+
+    if (isSelected) {
+        const filtered = selected.filter(currentId => currentId !== id)
+        return setSelected(filtered)
+    }
+
+    return setSelected([...selected, id])
+}
+
+export const createToggleAllHandler = ({
+    allSelected,
+    setSelected,
+    allIds,
+}) => () => {
+    if (allSelected) {
+        return setSelected([])
+    }
+
+    return setSelected(allIds)
+}
